requirements.txt
# Core dependencies for Kafka Ops Agent
flask>=2.3.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
aiosqlite>=0.19.0
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.8.0
kafka-python>=2.0.2
prometheus-client>=0.17.0
psutil>=5.9.0

# Optional dependencies
redis>=4.5.0
kubernetes>=27.0.0
docker>=6.1.0